v1.4 精确输入: 用户现在可以输入如120.50、85.25等精确的BPM值 范围限制: 支持1.00到999.99的BPM范围 显示格式: 所有显示都保持小数点后两位的格式 向后兼容: 整数BPM值仍然完全支持 精度保持: 在处理过程中保持高精度的时间计算

+++ midi_processor.py
@@ -15,7 +15,7 @@ class MidiProcessor:
     def process_file(self, 
                     input_file: str, 
                     output_dir: str, 
-                    target_bpm: int = 120, 
+                    target_bpm: float = 120.0, 
                     remove_cc: bool = True, 
                     set_velocity: bool = True,
                     velocity_percent: int = 80,
@@ -362,7 +362,7 @@ class MidiProcessor:
     def _create_new_midi_with_exact_timing(self, 
                                         orig_midi: mido.MidiFile, 
                                         note_positions: List[Dict[str, Any]],
-                                        target_bpm: int,
+                                        target_bpm: float,
                                         remove_cc: bool,
                                         set_velocity: bool) -> mido.MidiFile:
         """
@@ -789,7 +789,7 @@ class MidiProcessor:
     def process_directory(self, 
                          input_dir: str, 
                          output_dir: str, 
-                         target_bpm: int = 120, 
+                         target_bpm: float = 120.0, 
                          remove_cc: bool = True, 
                          set_velocity: bool = True,
                          velocity_percent: int = 80,
